requirements.txt
agithub~=2.2
pyYAML~=5.3
PyNaCl~=1.3.0
loguru~=0.4.1
